+++ libs/zhi-infra/src/index.spec.ts
@@ -1,0 +1,8 @@
+import { describe, expect, it } from "vitest"
+import init from "./index"
+
+describe("zhi-server-infra", () => {
+  it("index", () => {
+    expect(init()).toBe("ok")
+  })
+})