+++ libs/zhi-chatgpt/CHANGELOG.md
@@ -1,5 +1,12 @@
 # zhi-chatgpt
 
+## 0.0.3
+
+### Patch Changes
+
+- feat: add zhi-infra
+- add zhi-chatgpt zhi-infra
+
 ## 0.0.2
 
 ### Patch Changes